+++ include/invader/tag/parser/compile/model.hpp
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: GPL-3.0-only
+
+#ifndef INVADER__TAG__PARSER__COMPILE__MODEL_HPP
+#define INVADER__TAG__PARSER__COMPILE__MODEL_HPP
+
+namespace Invader::Parser {
+    struct GBXModel;
+    struct GBXModelGeometryPart;
+    bool uncache_model_markers(GBXModel &model, bool fix);
+    bool regenerate_missing_model_vertices(GBXModelGeometryPart &part, bool fix);
+    bool regenerate_missing_model_vertices(GBXModel &model, bool fix);
+    
+    struct Model;
+    struct ModelGeometryPart;
+    bool uncache_model_markers(Model &model, bool fix);
+    bool regenerate_missing_model_vertices(ModelGeometryPart &part, bool fix);
+    bool regenerate_missing_model_vertices(Model &model, bool fix);
+}
+
+#endif