+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This is used for recording Invader's changes. This changelog is based on
   queried
 - invader-edit-qt: Added a filter option in the View menu to filter out tags
   that don't match a given expression
+- invader-edit-qt: Added support for model tags (as opposed to gbxmodel)
 - invader-font: Added `-l` to limit to the first 256 characters
 - invader-resource: Added `-p` to specify a number of extra bytes to add after
   the header when generating resource maps